Bump spree chimpy to Solidus
Big spec changes: Preferences must be a hash, they cannot be nil.
After creating a user, you are redirected to users#edit instead of
rendering #edit on the create path. Transitioning to complete must be
explicit instead of using next. The controller requests module uses the
routes helper, that's only available in controller specs and thus must
be restricted to those.